+++ signer/src/dkg/mod.rs
@@ -1,0 +1,7 @@
+//! Module containing DKG-specific types and functions.
+
+pub mod verification;
+
+#[cfg(test)]
+mod testing;
+mod wsts;
